refactor: consolidate R2 scripts into single module with subcommands
Merge upload_r2.py and cleanup_r2.py into r2.py with:
- `r2.py upload` - upload site with content-addressed storage
- `r2.py cleanup` - clean up orphaned blobs after PR closure

Shared S3 client setup and credential handling are now deduplicated.
