+++ .github/workflows/cleanup-r2.yml
@@ -49,6 +49,6 @@ jobs:
           R2_ACCESS_KEY_ID: ${{ secrets.R2_ACCESS_KEY_ID }}
           R2_SECRET_ACCESS_KEY: ${{ secrets.R2_SECRET_ACCESS_KEY }}
         run: |
-          uv run python scripts/cleanup_r2.py \
+          uv run python scripts/r2.py cleanup \
             --closed-pr ${{ github.event.pull_request.number }} \
             --open-prs "${{ steps.open-prs.outputs.list }}"
