+++ .github/workflows/sync.yml
@@ -200,7 +200,7 @@ jobs:
           R2_ACCESS_KEY_ID: ${{ secrets.R2_ACCESS_KEY_ID }}
           R2_SECRET_ACCESS_KEY: ${{ secrets.R2_SECRET_ACCESS_KEY }}
         run: |
-          uv run python scripts/upload_r2.py \
+          uv run python scripts/r2.py upload \
             --dist site/dist \
             --manifest "${{ steps.manifest.outputs.name }}"
